+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added `mediable.allowed_remote_hosts` configuration which allows restricting remote URL source adapters to a whitelist. Wildcard subdomains may be specified with `*.example.com` syntax. By default, all hosts are allowed.
 * If no `mediable.allowed_remote_hosts` allow list is provided, the `RemoteUrlAdapter` will reject private IP addresses and localhost by default, to prevent Server-Side Request Forgery (SSRF) attacks. [CVE-2026-49969]
 * Added `mediable.allowed_remote_schemes` configuration which allows restricting remote URL source adapters to a whitelist of allowed URL schemes (e.g. `http`, `https`, `ftp`, etc.). By default, only `https` is allowed. When modifying this configuration, it may also be necessary to modify the `source_adapters` pattern matching.
+* Directory validation will now strip `.` characters from path segments. This prevents attackers from using directory traversal patterns like `../`  as well as preventing potential issues with certain filesystems where `.` characters in directory names may cause unexpected behaviour. [CVE-2026-49970]
 
 ## 6.1.0