+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 6.2.0
+## 7.0.0
 
 **This is a security release, upgrading is strongly recommended**
 * Added `mediable.allowed_remote_hosts` configuration which allows restricting remote URL source adapters to a whitelist. Wildcard subdomains may be specified with `*.example.com` syntax. By default, all hosts are allowed.
@@ -9,6 +9,8 @@
 * Directory validation will now strip `.` characters from path segments. This prevents attackers from using directory traversal patterns like `../`  as well as preventing potential issues with certain filesystems where `.` characters in directory names may cause unexpected behaviour. [CVE-2026-49970]
 * Added `mediable.file_sanitizers` configuration which allows specifying custom sanitizers for rewriting file contents to strip out security risks before they are uploaded. A sanitizer must implements the `Plank\Mediable\Sanitizers\SanitizerInterface` interface.
 * Added `SvgSanitizer` which will strip executable javascript and other untrusted content from `image/svg+xml` files, which can result in stored XSS if rendered directly to a webpage. By default, this sanitizer is applied to all uploaded SVG files. It can be disabled by removing it from the `mediable.sanitizers` config array. [CVE-2026-49971]
+* Added `mediable.forbidden_file_extensions` configuration and `MediaUploader::setForbiddenExtensions()` which allows specifying a blacklist of file extensions that are forbidden to be uploaded. Any file extension which is considered executable by your Apache or Nginx configuration should be included in this list. A number of common executable file extensions are included in this list by default to prevent remote code execution exploits.
+* Added `mediable.forbidden_mime_types` configuration and `MediaUploader::setForbiddenMimeTypes()` which allows specifying a blacklist of MIME types that are forbidden to be uploaded.
 
 ## 6.1.0